+++ docs/CHANGELOG.md
@@ -15,6 +15,13 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.26.0] - 29-10-23
+
+### Added
+- Added rules2excel rules exporter. Now users can export rules from TransformationRules object to excel file.
+- Added rules generator step from arbitrary object (must be in json or yaml formats)
+- Added eperimental rules parser from OpenApi/Swagger specification. Rules generates based on schema part of OpenApi specification.
+
 ## [0.25.8] - 20-09-23
 
 ### Improved